refactor: refactors Whitelist class into a Rails model
