+++ app/constraints/staff_only_constraint.rb
@@ -1,10 +1,8 @@
 # frozen_string_literal: true
 
-require "nla/blacklight/user_location"
-
 class StaffOnlyConstraint
   def matches?(request)
-    location = Nla::Blacklight::UserLocation.instance.location request
+    location = Whitelist.instance.location request
     location == :staff || Rails.env.development?
   end
 end
