+++ app/helpers/patron_helper.rb
@@ -1,14 +1,11 @@
 # frozen_string_literal: true
 
-require "nla/blacklight/user_location"
-require "nla/blacklight/user_type"
-
 module PatronHelper
   def user_location
-    Nla::Blacklight::UserLocation.instance.location request
+    Whitelist.instance.location request
   end
 
   def user_type
-    Nla::Blacklight::UserType.instance.type request
+    Whitelist.instance.user_type request
   end
 end
